Move request dto to new package
