+++ src/main/java/com/sfjs/conv/PaymentConverter.java
@@ -2,9 +2,9 @@ package com.sfjs.conv;
 
 import org.springframework.stereotype.Service;
 
-import com.sfjs.dto.BusinessRequest;
-import com.sfjs.dto.FellowRequest;
-import com.sfjs.dto.PaymentRequest;
+import com.sfjs.dto.request.BusinessRequest;
+import com.sfjs.dto.request.FellowRequest;
+import com.sfjs.dto.request.PaymentRequest;
 import com.sfjs.dto.response.PaymentResponse;
 import com.sfjs.entity.BusinessEntity;
 import com.sfjs.entity.FellowEntity;
