+++ src/main/java/com/sfjs/dto/request/BaseRequest.java
@@ -1,4 +1,4 @@
-package com.sfjs.dto;
+package com.sfjs.dto.request;
 
 import java.time.LocalDateTime;
