+++ src/main/java/com/sfjs/dto/request/BusinessRequest.java
@@ -1,4 +1,4 @@
-package com.sfjs.dto;
+package com.sfjs.dto.request;
 
 import lombok.Getter;
 import lombok.Setter;
