+++ src/main/java/com/sfjs/dto/request/PaymentRequest.java
@@ -1,4 +1,4 @@
-package com.sfjs.dto;
+package com.sfjs.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
