+++ src/main/java/com/sfjs/dto/request/RoleRequest.java
@@ -1,4 +1,4 @@
-package com.sfjs.dto;
+package com.sfjs.dto.request;
 
 public class RoleRequest extends BaseRequest {
