+++ src/main/java/com/sfjs/gql/svc/HelcimService.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sfjs.dto.PaymentRequest;
+import com.sfjs.dto.request.PaymentRequest;
 import com.sfjs.entity.PaymentEntity;
 
 import jakarta.annotation.PostConstruct;
