+++ src/main/java/com/sfjs/gql/svc/SignupService.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Service;
 
 import com.sfjs.conv.BusinessConverter;
 import com.sfjs.conv.FellowConverter;
-import com.sfjs.dto.BusinessRequest;
-import com.sfjs.dto.FellowRequest;
+import com.sfjs.dto.request.BusinessRequest;
+import com.sfjs.dto.request.FellowRequest;
 import com.sfjs.dto.response.BusinessResponse;
 import com.sfjs.dto.response.FellowResponse;
 import com.sfjs.entity.AccountEntity;
