+++ src/main/java/com/sfjs/rest/BusinessController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
-import com.sfjs.dto.BusinessRequest;
+import com.sfjs.dto.request.BusinessRequest;
 import com.sfjs.dto.response.BusinessResponse;
 import com.sfjs.entity.BusinessEntity;
 import com.sfjs.svc.BusinessService;
