+++ src/main/java/com/sfjs/rest/NumericMetricController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
-import com.sfjs.dto.NumericMetricRequest;
+import com.sfjs.dto.request.NumericMetricRequest;
 import com.sfjs.dto.response.NumericMetricResponse;
 import com.sfjs.entity.NumericMetricEntity;
 import com.sfjs.svc.NumericMetricService;
