+++ src/main/java/com/sfjs/svc/PaymentService.java
@@ -16,9 +16,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sfjs.conv.BusinessConverter;
 import com.sfjs.conv.FellowConverter;
 import com.sfjs.conv.PaymentConverter;
-import com.sfjs.dto.PaymentRequest;
 import com.sfjs.dto.PaymentResult;
 import com.sfjs.dto.PaymentResultInput;
+import com.sfjs.dto.request.PaymentRequest;
 import com.sfjs.dto.response.PaymentResponse;
 import com.sfjs.entity.NumericMetricEntity;
 import com.sfjs.entity.PaymentEntity;
