+++ src/main/java/com/sfjs/svc/RoleService.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.sfjs.conv.BaseConverter;
-import com.sfjs.dto.RoleRequest;
+import com.sfjs.dto.request.RoleRequest;
 import com.sfjs.dto.response.RoleResponse;
 import com.sfjs.entity.RoleEntity;
 import com.sfjs.repo.BaseRepository;
